+++ whylabs_toolkit/helpers/utils.py
@@ -1,11 +1,8 @@
-from typing import Union
-
 from whylabs_client.api.dataset_profile_api import DatasetProfileApi
 from whylabs_client.api.models_api import ModelsApi
 from whylabs_client.api.notification_settings_api import NotificationSettingsApi
 
 from whylabs_toolkit.helpers.client import create_client
-from whylabs_toolkit.monitor.models import SlackWebhook, EmailRecipient
 
 
 def get_models_api() -> ModelsApi:
@@ -18,12 +15,3 @@ def get_dataset_profile_api() -> DatasetProfileApi:
 
 def get_notification_api() -> NotificationSettingsApi:
     return NotificationSettingsApi(api_client=create_client())
-
-
-def get_notification_request_payload(action: Union[SlackWebhook, EmailRecipient]) -> str:
-    if isinstance(action, SlackWebhook):
-        return "slackWebhook"
-    elif isinstance(action, EmailRecipient):
-        return "email"
-    else:
-        raise ValueError(f"Can't work with {action} type. Available options are SlackWebhook and EmailRecipient.")
