+++ whylabs_toolkit/monitor/manager/monitor_setup.py
@@ -2,14 +2,14 @@ import pytz
 import logging
 from datetime import datetime
 from abc import abstractmethod
+from typing import Optional, List, Union, Any
 
 from whylabs_client.exceptions import NotFoundException
 
-from whylabs_toolkit.helpers.monitor_helpers import get_analyzers, get_monitor
 from whylabs_toolkit.helpers.utils import get_models_api
 from whylabs_toolkit.monitor.models import *
-from whylabs_toolkit.monitor.models.analyzer.algorithms import *
 from whylabs_toolkit.monitor.manager.credentials import MonitorCredentials
+from whylabs_toolkit.helpers.monitor_helpers import get_analyzers, get_monitor
 
 
 logging.basicConfig(level=logging.INFO)
